requirements.txt
chromadb
agentlogger
psycopg2-binary
sentence_transformers
python-dotenv